User Select Solo or Multiplayer
Created a page that allows the user to select solo or multiplayer games and goes to that page. Multiplayer is working as intended, Solo still does not have a cpu yet
