+++ src/App.js
@@ -1,71 +1,26 @@
 // App.js
 import React, { useState } from 'react';
 import { Client } from 'boardgame.io/react';
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { SocketIO } from 'boardgame.io/multiplayer';
 import { headerStyle, mainStyle } from './Data/inlineStyle.js';
 import { ConnectFour } from './Game/Game.js';
 import { ConnectFourBoard } from './Game/Board.js';
 import LobbySetup from './LobbySetup';
-
-const ConnectFourClient = Client({
-  game: ConnectFour,
-  board: ConnectFourBoard,
-  multiplayer: SocketIO({ server: 'http://localhost:8000' }),
-});
+import MultiplayerPage from './Pages/MultiplayerPage.jsx';
+import Home from './Pages/Home.jsx';
+import SoloPage from './Pages/SoloPage.jsx';
 
 const App = () => {
-  const [playerName, setPlayerName] = useState(null);
-  const [playerID, setPlayerID] = useState(null);
-  const [matchID, setMatchID] = useState(null);
-
-  const onMatchJoined = (name, matchID, isCreator) => {
-    setPlayerName(name);
-    setMatchID(matchID);
-    // Assign player ID based on whether the player is creating or joining
-    setPlayerID(isCreator ? '0' : '1');
-  };
-
   return (
-    <div style={fullDisplay}>
-      <div className="header" style={headerStyle}>
-        <p>Connect Four Online</p>
-      </div>
-
-      <div style={mainStyle}>
-        {!playerID ? (
-          <LobbySetup onMatchJoined={onMatchJoined} />
-        ) : (
-          <div style={boardFlexStyle}>
-            <ConnectFourClient
-              playerID={playerID}
-              matchID={matchID}
-              credentials={playerName}
-            />
-          </div>
-        )}
-
-        <div style={chatFlexStyle}>
-          {/* Chat functionality or additional components can be added here */}
-        </div>
-      </div>
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/solo" element={<SoloPage />} />
+        <Route path="/multiplayer" element={<MultiplayerPage />} />
+        <Route path="/" element={<Home />} />
+      </Routes>
+    </Router>
   );
-};
-
-const fullDisplay = {
-  minHeight: '100vh',
-};
-
-const boardFlexStyle = {
-  flexBasis: '75%',
-  justifyContent: 'center',
-  display: 'flex',
-  alignItems: 'center',
-};
-
-const chatFlexStyle = {
-  flexBasis: '25%',
-  background: 'lightgrey',
 };
 
 export default App;
